It now longer throws big bad errors around.. but the screen is still empty, my guess is that its a big bad pointer somewhere.

+++ jni/render.h
@@ -29,7 +29,7 @@ typedef struct {
 	GLfloat* textureCoordBuffer;
 	GLsizeiptr indexBufSize;
 	GLuint* indexBuffer;
-	GLuint 	indexCount;
+	GLushort 	indexCount;
 	
 	GLuint bufferName[3];
